Did a lot of work to get notifications working as well as all the ways of sending, receiving, accepting, denying, and removing friend requests, access requests, and publisher/literary agent representation of writers.
I'm currently in the process of making it so that there is the ability to restrict access to only publishers or literary agents who represent you. The updates have been made in the access controller, but there still need to be some updates made on the view components, the AccessDenied views, and the AccessPermission sections of the create and edit for profile and writing.